pyproject.toml
[project]
name = "mcp-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]>=1.10.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]
